+++ dynamic_programming/maximum_subarray/subarray.h
@@ -1,6 +1,8 @@
 #ifndef SUBARRAY_H_
 #define SUBARRAY_H_
 
-int max_subarray(int *, int);
+#include <stddef.h>
+
+int max_subarray(int *, size_t);
 
 #endif
